bugfix
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@20695 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_fluid_xfluid/xfluid_outputservice.cpp
@@ -68,7 +68,7 @@ void FLD::XFluidOutputService::PrepareOutput()
 
 void FLD::XFluidOutputService::Output(
   int step,
-  int time,
+  double time,
   bool write_restart_data,
   Teuchos::RCP<const FLD::XFluidState> state,
   Teuchos::RCP<const Epetra_Vector>    dispnp,
